Add page to explain role permissions, user mgmt
Hello! I'm new to this, so forgive me if this the incorrect way to contribute.

I wanted to add a page that explained the user permissions because I couldn't find this info anywhere. The tables were largely inspired from this page: https://support.airtable.com/hc/en-us/articles/202887099-Permissions-overview.

Please complete the table with check-marks and x's to indicate what the various roles can or can't do.